Updated README package purpose
Previously the mission statement was as follows:

> `report`'s primary goal is to fill the gap between R's output and the formatted result description of your manuscript, with the automated use of **best practices** guidelines (*e.g.,* [APA](https://www.apastyle.org/)'s style guide), ensuring **standardization** and **quality** of results reporting.

There were too many clauses that I think naive individuals (who this package is targeted at) are unlikely to quickly grasp its purpose, and there were grammatical issues that caused a lack of clarity. I hope the suggested changes here are appropriate and resolve the issues.

The changes were adusted for both the Rmd and md README files.